feat: auth 경로 변경 및 로그아웃 플로우 변경
